+++ examples/hosts/app-integration/external-views/package.json
@@ -44,8 +44,6 @@
     "@fluidframework/map": "^0.23.0",
     "@fluidframework/protocol-definitions": "^0.1009.0",
     "@fluidframework/routerlicious-driver": "^0.23.0",
-    "@fluidframework/view-adapters": "^0.23.0",
-    "@fluidframework/view-interfaces": "^0.23.0",
     "@fluidframework/web-code-loader": "^0.23.0",
     "jsonwebtoken": "^8.4.0",
     "react": "^16.10.2",
