+++ examples/hosts/app-integration/external-views/src/app/views/index.ts
@@ -3,5 +3,4 @@
  * Licensed under the MIT License.
  */
 
-export * from "./interface";
-export * from "./model";
+export * from "./PrettyDiceRollerView";
